4960: Move ingest batcher submitter to single module
