+++ .github/config/chunks.yaml
@@ -77,7 +77,6 @@ chunks:
       - ingest/ingest-batcher-core
       - ingest/ingest-batcher-store
       - ingest/ingest-batcher-submitter
-      - ingest/ingest-batcher-submitterv2
       - ingest/ingest-batcher-job-creator
       - bulk-import/bulk-import-core
       - bulk-import/bulk-import-starter
