+++ .github/workflows/chunk-ingest.yaml
@@ -17,7 +17,6 @@ on:
       - 'java/ingest/ingest-batcher-core/**'
       - 'java/ingest/ingest-batcher-store/**'
       - 'java/ingest/ingest-batcher-submitter/**'
-      - 'java/ingest/ingest-batcher-submitterv2/**'
       - 'java/ingest/ingest-batcher-job-creator/**'
       - 'java/bulk-import/bulk-import-core/**'
       - 'java/bulk-import/bulk-import-starter/**'
