+++ java/ingest/ingest-batcher-submitter/src/main/java/sleeper/ingest/batcher/submitter/IngestBatcherSubmitterLambda.java
@@ -15,23 +15,19 @@
  */
 package sleeper.ingest.batcher.submitter;
 
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.SQSEvent;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.amazonaws.services.sqs.AmazonSQS;
-import com.amazonaws.services.sqs.AmazonSQSClientBuilder;
-import org.apache.hadoop.conf.Configuration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.sqs.SqsClient;
 
-import sleeper.configuration.properties.S3InstanceProperties;
-import sleeper.configuration.properties.S3PropertiesReloader;
-import sleeper.configuration.properties.S3TableProperties;
-import sleeper.configuration.table.index.DynamoDBTableIndex;
+import sleeper.configurationv2.properties.S3InstanceProperties;
+import sleeper.configurationv2.properties.S3PropertiesReloader;
+import sleeper.configurationv2.properties.S3TableProperties;
+import sleeper.configurationv2.table.index.DynamoDBTableIndex;
 import sleeper.core.properties.PropertiesReloader;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TablePropertiesProvider;
@@ -40,7 +36,6 @@ import sleeper.ingest.batcher.core.IngestBatcherStore;
 import sleeper.ingest.batcher.core.IngestBatcherSubmitRequest;
 import sleeper.ingest.batcher.core.IngestBatcherSubmitRequestSerDe;
 import sleeper.ingest.batcher.store.DynamoDBIngestBatcherStore;
-import sleeper.parquet.utils.HadoopConfigurationProvider;
 
 import java.time.Instant;
 
@@ -58,28 +53,26 @@ public class IngestBatcherSubmitterLambda implements RequestHandler<SQSEvent, Vo
         if (null == s3Bucket) {
             throw new IllegalArgumentException("Couldn't get S3 bucket from environment variable");
         }
-        AmazonS3 s3Client = AmazonS3ClientBuilder.defaultClient();
-        AmazonSQS sqsClient = AmazonSQSClientBuilder.defaultClient();
-        AmazonDynamoDB dynamoDBClient = AmazonDynamoDBClientBuilder.defaultClient();
+        S3Client s3Client = S3Client.create();
+        SqsClient sqsClient = SqsClient.create();
+        DynamoDbClient dynamoDBClient = DynamoDbClient.create();
 
         InstanceProperties instanceProperties = S3InstanceProperties.loadFromBucket(s3Client, s3Bucket);
 
         TablePropertiesProvider tablePropertiesProvider = S3TableProperties.createProvider(instanceProperties, s3Client, dynamoDBClient);
         this.propertiesReloader = S3PropertiesReloader.ifConfigured(s3Client, instanceProperties, tablePropertiesProvider);
         this.deadLetterQueue = new IngestBatcherSubmitDeadLetterQueue(instanceProperties, sqsClient);
-        this.submitter = new IngestBatcherSubmitter(instanceProperties,
-                HadoopConfigurationProvider.getConfigurationForLambdas(instanceProperties),
-                new DynamoDBTableIndex(instanceProperties, dynamoDBClient),
+        this.submitter = new IngestBatcherSubmitter(new DynamoDBTableIndex(instanceProperties, dynamoDBClient),
                 new DynamoDBIngestBatcherStore(dynamoDBClient, instanceProperties, tablePropertiesProvider),
-                deadLetterQueue);
+                deadLetterQueue, s3Client);
     }
 
     public IngestBatcherSubmitterLambda(
             IngestBatcherStore store, InstanceProperties instanceProperties,
-            TableIndex tableIndex, Configuration conf, IngestBatcherSubmitDeadLetterQueue dlQueue) {
+            TableIndex tableIndex, IngestBatcherSubmitDeadLetterQueue dlQueue, S3Client s3Client) {
         this.propertiesReloader = PropertiesReloader.neverReload();
         this.deadLetterQueue = dlQueue;
-        this.submitter = new IngestBatcherSubmitter(instanceProperties, conf, tableIndex, store, deadLetterQueue);
+        this.submitter = new IngestBatcherSubmitter(tableIndex, store, deadLetterQueue, s3Client);
     }
 
     @Override
