+++ java/ingest/ingest-batcher-submitter/src/test/java/sleeper/ingest/batcher/submitter/IngestBatcherSubmitterLambdaIT.java
@@ -16,22 +16,24 @@
 
 package sleeper.ingest.batcher.submitter;
 
-import com.amazonaws.services.sqs.model.Message;
-import com.amazonaws.services.sqs.model.ReceiveMessageRequest;
-import com.amazonaws.services.sqs.model.ReceiveMessageResult;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import software.amazon.awssdk.services.sqs.model.Message;
+import software.amazon.awssdk.services.sqs.model.ReceiveMessageRequest;
+import software.amazon.awssdk.services.sqs.model.ReceiveMessageResponse;
 
 import sleeper.core.properties.instance.InstanceProperties;
+import sleeper.core.properties.table.TableProperties;
+import sleeper.core.properties.testutils.FixedTablePropertiesProvider;
 import sleeper.core.table.InMemoryTableIndex;
 import sleeper.core.table.TableIndex;
 import sleeper.core.table.TableStatusTestHelper;
 import sleeper.ingest.batcher.core.IngestBatcherStore;
 import sleeper.ingest.batcher.core.IngestBatcherTrackedFile;
-import sleeper.ingest.batcher.core.testutil.InMemoryIngestBatcherStore;
+import sleeper.ingest.batcher.store.DynamoDBIngestBatcherStore;
+import sleeper.ingest.batcher.store.DynamoDBIngestBatcherStoreCreator;
 import sleeper.localstack.test.LocalStackTestBase;
 
 import java.time.Instant;
@@ -41,31 +43,34 @@ import java.util.UUID;
 import static net.javacrumbs.jsonunit.assertj.JsonAssertions.assertThatJson;
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.INGEST_BATCHER_SUBMIT_DLQ_URL;
+import static sleeper.core.properties.table.TableProperty.TABLE_ID;
+import static sleeper.core.properties.table.TableProperty.TABLE_NAME;
 import static sleeper.core.properties.testutils.InstancePropertiesTestHelper.createTestInstanceProperties;
+import static sleeper.core.properties.testutils.TablePropertiesTestHelper.createTestTableProperties;
+import static sleeper.core.schema.SchemaTestHelper.createSchemaWithKey;
 
 public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
 
     private static final String TEST_TABLE_ID = "test-table-id";
     private static final Instant RECEIVED_TIME = Instant.parse("2023-06-16T10:57:00Z");
-    private final ObjectMapper mapper = new ObjectMapper();
     private final String testBucket = UUID.randomUUID().toString();
-    private final IngestBatcherStore store = new InMemoryIngestBatcherStore();
     private final InstanceProperties instanceProperties = createTestInstanceProperties();
     private final TableIndex tableIndex = new InMemoryTableIndex();
-    private final IngestBatcherSubmitDeadLetterQueue dlQueue = new IngestBatcherSubmitDeadLetterQueue(instanceProperties, sqsClient);
-    private final IngestBatcherSubmitterLambda lambda = new IngestBatcherSubmitterLambda(
-            store, instanceProperties, tableIndex, hadoopConf, dlQueue);
+    private final TableProperties tableProperties = createTestTableProperties(instanceProperties, createSchemaWithKey("key"));
 
     @BeforeEach
     void setup() {
         tableIndex.create(TableStatusTestHelper.uniqueIdAndName(TEST_TABLE_ID, "test-table"));
         createBucket(testBucket);
+        DynamoDBIngestBatcherStoreCreator.create(instanceProperties, dynamoClientV2);
         instanceProperties.set(INGEST_BATCHER_SUBMIT_DLQ_URL, createSqsQueueGetUrl());
+        tableProperties.set(TABLE_ID, TEST_TABLE_ID);
+        tableProperties.set(TABLE_NAME, "test-table");
     }
 
     @Nested
-    @DisplayName("Store single file")
-    class StoreSingleFile {
+    @DisplayName("Store files by their full path")
+    class StoreFileByFullPath {
         @Test
         void shouldStoreFileIngestRequestFromJson() {
             // Given
@@ -76,12 +81,50 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst())
+            assertThat(batcherStore().getAllFilesNewestFirst())
                     .containsExactly(
                             fileRequest(testBucket + "/test-file-1.parquet"));
+            assertThat(receiveDeadLetters()).isEmpty();
+        }
+
+        @Test
+        void shouldStoreFileByPathNotIncludingItsExtension() {
+            // Given
+            uploadFileToS3("test-file-1.parquet");
+            String json = "{" +
+                    "\"files\":[\"" + testBucket + "/test-file-1\"]," +
+                    "\"tableName\":\"test-table\"" +
+                    "}";
+
+            // When
+            lambda().handleMessage(json, RECEIVED_TIME);
+
+            // Then
+            assertThat(batcherStore().getAllFilesNewestFirst())
+                    .containsExactly(
+                            fileRequest(testBucket + "/test-file-1.parquet"));
+            assertThat(receiveDeadLetters()).isEmpty();
+        }
+
+        @Test
+        void shouldStoreFileThatHasNoExtension() {
+            // Given
+            uploadFileToS3("test-file-1");
+            String json = "{" +
+                    "\"files\":[\"" + testBucket + "/test-file-1\"]," +
+                    "\"tableName\":\"test-table\"" +
+                    "}";
+
+            // When
+            lambda().handleMessage(json, RECEIVED_TIME);
+
+            // Then
+            assertThat(batcherStore().getAllFilesNewestFirst())
+                    .containsExactly(
+                            fileRequest(testBucket + "/test-file-1"));
             assertThat(receiveDeadLetters()).isEmpty();
         }
 
@@ -95,10 +138,10 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst())
+            assertThat(batcherStore().getAllFilesNewestFirst())
                     .containsExactly(IngestBatcherTrackedFile.builder()
                             .file(testBucket + "/test-file-1.parquet")
                             .fileSizeBytes(123)
@@ -106,11 +149,47 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                             .receivedTime(RECEIVED_TIME)
                             .build());
         }
+
+        @Test
+        void shouldStoreFileInDirectory() {
+            // Given
+            uploadFileToS3("test-directory/test-file-1.parquet");
+            String json = "{" +
+                    "\"files\":[\"" + testBucket + "/test-directory/test-file-1.parquet\"]," +
+                    "\"tableName\":\"test-table\"" +
+                    "}";
+
+            // When
+            lambda().handleMessage(json, RECEIVED_TIME);
+
+            // Then
+            assertThat(batcherStore().getAllFilesNewestFirst())
+                    .containsExactly(
+                            fileRequest(testBucket + "/test-directory/test-file-1.parquet"));
+        }
+
+        @Test
+        void shouldStoreFileInNestedDirectory() {
+            // Given
+            uploadFileToS3("test-directory/another-test-directory/test-file-1.parquet");
+            String json = "{" +
+                    "\"files\":[\"" + testBucket + "/test-directory/another-test-directory/test-file-1.parquet\"]," +
+                    "\"tableName\":\"test-table\"" +
+                    "}";
+
+            // When
+            lambda().handleMessage(json, RECEIVED_TIME);
+
+            // Then
+            assertThat(batcherStore().getAllFilesNewestFirst())
+                    .containsExactly(
+                            fileRequest(testBucket + "/test-directory/another-test-directory/test-file-1.parquet"));
+        }
     }
 
     @Nested
-    @DisplayName("Store all files in directory")
-    class StoreFilesInDirectory {
+    @DisplayName("Store files by a prefix/directory")
+    class StoreFilesByPrefix {
         @Test
         void shouldStoreOneFileInDirectory() {
             // Given
@@ -121,10 +200,10 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst())
+            assertThat(batcherStore().getAllFilesNewestFirst())
                     .containsExactly(
                             fileRequest(testBucket + "/test-directory/test-file-1.parquet"));
         }
@@ -140,17 +219,35 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst())
+            assertThat(batcherStore().getAllFilesNewestFirst())
                     .containsExactly(
                             fileRequest(testBucket + "/test-directory/test-file-1.parquet"),
                             fileRequest(testBucket + "/test-directory/test-file-2.parquet"));
         }
 
         @Test
-        void shouldStoreFileInNestedDirectories() {
+        void shouldStoreFileByNestedDirectory() {
+            // Given
+            uploadFileToS3("test-directory/another-test-directory/test-file-1.parquet");
+            String json = "{" +
+                    "\"files\":[\"" + testBucket + "/test-directory/another-test-directory\"]," +
+                    "\"tableName\":\"test-table\"" +
+                    "}";
+
+            // When
+            lambda().handleMessage(json, RECEIVED_TIME);
+
+            // Then
+            assertThat(batcherStore().getAllFilesNewestFirst())
+                    .containsExactly(
+                            fileRequest(testBucket + "/test-directory/another-test-directory/test-file-1.parquet"));
+        }
+
+        @Test
+        void shouldStoreFileInNestedDirectory() {
             // Given
             uploadFileToS3("test-directory/nested/test-file-1.parquet");
             String json = "{" +
@@ -159,10 +256,10 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst())
+            assertThat(batcherStore().getAllFilesNewestFirst())
                     .containsExactly(
                             fileRequest(testBucket + "/test-directory/nested/test-file-1.parquet"));
         }
@@ -178,10 +275,10 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst())
+            assertThat(batcherStore().getAllFilesNewestFirst())
                     .containsExactly(
                             fileRequest(testBucket + "/test-directory/nested-1/test-file-1.parquet"),
                             fileRequest(testBucket + "/test-directory/nested-2/test-file-2.parquet"));
@@ -198,10 +295,10 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst())
+            assertThat(batcherStore().getAllFilesNewestFirst())
                     .containsExactly(
                             fileRequest(testBucket + "/test-file-1.parquet"),
                             fileRequest(testBucket + "/test-file-2.parquet"));
@@ -225,10 +322,10 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst())
+            assertThat(batcherStore().getAllFilesNewestFirst())
                     .containsExactly(
                             fileRequest(testBucket + "/test-file-1.parquet"),
                             fileRequest(testBucket + "/test-file-2.parquet"));
@@ -247,10 +344,10 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst())
+            assertThat(batcherStore().getAllFilesNewestFirst())
                     .isEmpty();
         }
     }
@@ -265,10 +362,10 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
             String json = "{";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst()).isEmpty();
+            assertThat(batcherStore().getAllFilesNewestFirst()).isEmpty();
             assertThat(receiveDeadLetters()).singleElement().isEqualTo(json);
         }
 
@@ -281,10 +378,10 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst()).isEmpty();
+            assertThat(batcherStore().getAllFilesNewestFirst()).isEmpty();
             assertThat(receiveDeadLetters())
                     .singleElement()
                     .satisfies(deadLetter -> assertThatJson(deadLetter).isEqualTo(json));
@@ -299,10 +396,28 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst()).isEmpty();
+            assertThat(batcherStore().getAllFilesNewestFirst()).isEmpty();
+            assertThat(receiveDeadLetters())
+                    .singleElement()
+                    .satisfies(deadLetter -> assertThatJson(deadLetter).isEqualTo(json));
+        }
+
+        @Test
+        void shouldLogMessageIfDirectoryDoesNotExistAndSendToDeadLetterQueue() {
+            // Given
+            String json = "{" +
+                    "\"files\":[\"" + testBucket + "/not-exists\"]," +
+                    "\"tableName\":\"test-table\"" +
+                    "}";
+
+            // When
+            lambda().handleMessage(json, RECEIVED_TIME);
+
+            // Then
+            assertThat(batcherStore().getAllFilesNewestFirst()).isEmpty();
             assertThat(receiveDeadLetters())
                     .singleElement()
                     .satisfies(deadLetter -> assertThatJson(deadLetter).isEqualTo(json));
@@ -317,10 +432,10 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
                     "}";
 
             // When
-            lambda.handleMessage(json, RECEIVED_TIME);
+            lambda().handleMessage(json, RECEIVED_TIME);
 
             // Then
-            assertThat(store.getAllFilesNewestFirst()).isEmpty();
+            assertThat(batcherStore().getAllFilesNewestFirst()).isEmpty();
             assertThat(receiveDeadLetters()).isEmpty();
         }
     }
@@ -338,13 +453,25 @@ public class IngestBatcherSubmitterLambdaIT extends LocalStackTestBase {
     }
 
     private List<String> receiveDeadLetters() {
-        ReceiveMessageResult result = sqsClient.receiveMessage(new ReceiveMessageRequest()
-                .withQueueUrl(instanceProperties.get(INGEST_BATCHER_SUBMIT_DLQ_URL))
-                .withMaxNumberOfMessages(10)
-                .withWaitTimeSeconds(1));
-        return result.getMessages().stream()
-                .map(Message::getBody)
+        ReceiveMessageResponse result = sqsClientV2.receiveMessage(ReceiveMessageRequest.builder()
+                .queueUrl(instanceProperties.get(INGEST_BATCHER_SUBMIT_DLQ_URL))
+                .maxNumberOfMessages(10)
+                .waitTimeSeconds(1)
+                .build());
+        return result.messages().stream()
+                .map(Message::body)
                 .toList();
     }
 
+    private IngestBatcherSubmitterLambda lambda() {
+        return new IngestBatcherSubmitterLambda(
+                batcherStore(), instanceProperties, tableIndex,
+                new IngestBatcherSubmitDeadLetterQueue(instanceProperties, sqsClientV2),
+                s3ClientV2);
+    }
+
+    private IngestBatcherStore batcherStore() {
+        return new DynamoDBIngestBatcherStore(dynamoClientV2, instanceProperties,
+                new FixedTablePropertiesProvider(tableProperties));
+    }
 }
