+++ java/ingest/ingest-batcher-submitter/src/test/resources/log4j.properties
