+++ java/ingest/pom.xml
@@ -34,7 +34,6 @@
         <module>ingest-trackerV2</module>
         <module>ingest-batcher-core</module>
         <module>ingest-batcher-submitter</module>
-        <module>ingest-batcher-submitterv2</module>
         <module>ingest-batcher-job-creator</module>
     </modules>
 </project>
